Feat/auth

+++ views/MBCommodityView.py
@@ -1,15 +1,13 @@
-from re import I
-from this import d
 from bson import ObjectId
 from bson.errors import InvalidId
 from sanic import Blueprint, json
 from sanic.response import Request
+from sanic_jwt import inject_user, scoped
 from sanic_openapi import doc
+from controllers.MBAuthScope import MBAuthScope
 from models.MBCommodity import MBCommodity
-
 from models.MBItem import MBItem
 from controllers.MBRequest import MBRequest
-from controllers import MBUtil
 from models.MBUser import MBUser
 
 commodity_api = Blueprint("Commodities", url_prefix="/commodity/")
@@ -20,7 +18,6 @@ commodity_api = Blueprint("Commodities", url_prefix="/commodity/")
     doc.String(name="item_id", description="Commodity item"),
     doc.Float(name="quandity", description="Commodity quandity"),
     doc.Float(name="price", description="Commodity price"),
-    doc.String(name="user_id", description="Commodity user owner"),
     required=True,
 )
 @doc.consumes(
@@ -28,9 +25,11 @@ commodity_api = Blueprint("Commodities", url_prefix="/commodity/")
     required=False,
 )
 @commodity_api.put("edit")
-async def edit_commodity(request: Request):
+@inject_user()
+@scoped(MBAuthScope.ADMIN, require_all=False)
+async def edit_commodity(request: Request, user: MBUser):
 
-    required = ["item_id", "quandity", "price", "user_id"]
+    required = ["item_id", "quandity", "price"]
     for param in required:
         if param not in request.args:
             return MBRequest.response_invalid_params([param])
@@ -57,14 +56,6 @@ async def edit_commodity(request: Request):
         except InvalidId:
             return MBRequest.invalid_item_id(item_id)
 
-    user_id = request.args.get("user_id")
-    user_object_id: ObjectId
-    if user_id is not None:
-        try:
-            user_object_id = ObjectId(user_id)
-        except InvalidId:
-            return MBRequest.invalid_user_id(user_id)
-
     commodity_id = request.args.get("commodity_id")
     commodity: MBCommodity = None
 
@@ -81,17 +72,14 @@ async def edit_commodity(request: Request):
         if commodity is None:
             return MBRequest.invalid_commodity_id(commodity_id)
 
-    user = await MBUser.find_one(user_object_id)
-
-    if user is None:
-        return MBRequest.invalid_user_id(user_id)
-
     item = await MBItem.find_one(item_object_id)
     if item is None:
         return MBRequest.invalid_item_id(item_id)
 
     if commodity is None:
-        commodity = await MBCommodity.create_commodity(item.id, quandity, price, user.id)
+        commodity = await MBCommodity.create_commodity(
+            item.id, quandity, price, user.id
+        )
     else:
         await MBCommodity.edit_commodity(commodity, item.id, quandity, price, user.id)
 
@@ -105,14 +93,15 @@ async def edit_commodity(request: Request):
     required=True,
 )
 @commodity_api.get("/<commodity_id>")
-async def get_item(request: Request, commodity_id: str):
-
+@inject_user()
+@scoped(MBAuthScope.USER, require_all=False)
+async def get_item(request: Request, commodity_id: str, user: MBUser):
     try:
         commodity_object_id = ObjectId(commodity_id)
     except InvalidId:
         return MBRequest.invalid_commodity_id(commodity_id)
 
-    commodity = await MBCommodity.find_one({"id": commodity_object_id})
+    commodity = await MBCommodity.find_one({"id": commodity_object_id, "user_id_own": user.id})
     if commodity is None:
         return MBRequest.invalid_commodity_id(commodity_id)
 
@@ -120,31 +109,16 @@ async def get_item(request: Request, commodity_id: str):
 
 
 @doc.summary("Get all commodities")
-@doc.consumes(doc.String(name="user_id", description="User owner"), required=True)
 @commodity_api.get("list")
-async def list_items(request: Request):
-
-    user_id = request.args.get("user_id")
-    user_object_id: ObjectId
-    if user_id is not None:
-        try:
-            user_object_id = ObjectId(user_id)
-        except InvalidId:
-            return MBRequest.invalid_user_id(user_id)
-    else:
-        return MBRequest.response_invalid_params(["user_id"])
-
-    user = await MBUser.find_one(user_object_id)
-
-    if user is None:
-        return MBRequest.invalid_user_id(user_id)
-
+@inject_user()
+@scoped(MBAuthScope.USER, require_all=False)
+async def list_items(request: Request, user: MBUser):
     commodities = []
     async for commodity in MBCommodity.find({"user_id_own": user.id}):
         commodity_object = MBCommodity.create_object_from_dict(commodity)
         commodities.append(await MBCommodity.to_api(commodity_object))
-
     return json({"commodities": commodities, "user": await MBUser.to_api(user)})
+
 
 @doc.summary("Remove commodity by id")
 @doc.consumes(
@@ -153,16 +127,23 @@ async def list_items(request: Request):
     required=True,
 )
 @commodity_api.delete("/<commodity_id>")
-async def remove_commodity(request: Request, commodity_id: str):
+@inject_user()
+@scoped(MBAuthScope.USER, require_all=False)
+async def remove_commodity(request: Request, commodity_id: str, user: MBUser):
     try:
         commodity_object_id = ObjectId(commodity_id)
     except InvalidId:
         return MBRequest.invalid_commodity_id(commodity_id)
 
-    commodity = await MBCommodity.find_one({"id": commodity_object_id})
-    await commodity.delete()
+    commodity = await MBCommodity.find_one(
+        {"id": commodity_object_id, "user_id_own": user.id}
+    )
+    if commodity is None:
+        return MBRequest.invalid_commodity_id(commodity_id)
 
+    await commodity.delete()
     return json({"deleted": commodity_id})
+
 
 @doc.summary("Sell commodity by id")
 @doc.consumes(
@@ -171,15 +152,18 @@ async def remove_commodity(request: Request, commodity_id: str):
     required=True,
 )
 @commodity_api.post("/<commodity_id>")
-async def sell_commodity(request: Request, commodity_id: str):
-
+@inject_user()
+@scoped(MBAuthScope.USER, require_all=False)
+async def sell_commodity(request: Request, commodity_id: str, user: MBUser):
     try:
         commodity_object_id = ObjectId(commodity_id)
     except InvalidId:
         return MBRequest.invalid_commodity_id(commodity_id)
-    
-    commodity = await MBCommodity.find_one({"id": commodity_object_id})
+
+    commodity = await MBCommodity.find_one(
+        {"id": commodity_object_id, "user_id_own": user.id}
+    )
     if commodity is None:
         return MBRequest.invalid_commodity_id(commodity_id)
-    
+
     return json(await MBCommodity.sell_commodity(commodity))
